Update wear module - remove complication service - refactored packages - fix language
