+++ instances/wearApp/src/main/AndroidManifest.xml
@@ -13,24 +13,6 @@
         android:supportsRtl="true"
         android:theme="@android:style/Theme.DeviceDefault">
         <service
-            android:name=".wear.complication.MainComplicationService"
-            android:exported="true"
-            android:icon="@mipmap/ic_launcher"
-            android:label="MainComplicationService"
-            android:permission="com.google.android.wearable.permission.BIND_COMPLICATION_PROVIDER"
-            android:roundIcon="@mipmap/ic_launcher">
-            <intent-filter>
-                <action android:name="android.support.wearable.complications.ACTION_COMPLICATION_UPDATE_REQUEST" />
-            </intent-filter>
-
-            <meta-data
-                android:name="android.support.wearable.complications.SUPPORTED_TYPES"
-                android:value="SHORT_TEXT" />
-            <meta-data
-                android:name="android.support.wearable.complications.UPDATE_PERIOD_SECONDS"
-                android:value="0" />
-        </service>
-        <service
             android:name=".wear.tile.StatusesTileService"
             android:exported="true"
             android:icon="@mipmap/ic_launcher"
