+++ instances/wearApp/src/main/java/com/makeevrserg/empireprojekt/mobile/wear/MainActivity.kt
@@ -8,8 +8,8 @@ import com.arkivanov.decompose.defaultComponentContext
 import com.makeevrserg.empireprojekt.mobile.core.resources.R
 import com.makeevrserg.empireprojekt.mobile.features.theme.ApplicationTheme
 import com.makeevrserg.empireprojekt.mobile.wear.application.App.Companion.asEmpireApp
-import com.makeevrserg.empireprojekt.mobile.wear.features.root.DefaultRootComponent
-import com.makeevrserg.empireprojekt.mobile.wear.features.root.RootScreen
+import com.makeevrserg.empireprojekt.mobile.wear.features.root.presentation.DefaultRootComponent
+import com.makeevrserg.empireprojekt.mobile.wear.features.root.ui.RootScreen
 import ru.astrainteractive.klibs.kdi.Provider
 import ru.astrainteractive.klibs.kdi.getValue
