+++ instances/wearApp/src/main/java/com/makeevrserg/empireprojekt/mobile/wear/features/root/ui/RootScreen.kt
@@ -1,18 +1,16 @@
-package com.makeevrserg.empireprojekt.mobile.wear.features.root
+package com.makeevrserg.empireprojekt.mobile.wear.features.root.ui
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import com.makeevrserg.empireprojekt.mobile.wear.features.main.MainScreen
+import com.makeevrserg.empireprojekt.mobile.wear.features.main.ui.MainScreen
 import com.makeevrserg.empireprojekt.mobile.wear.features.ping.ui.PingScreen
-import com.makeevrserg.empireprojekt.mobile.wear.features.status.StatusesScreen
+import com.makeevrserg.empireprojekt.mobile.wear.features.root.presentation.RootComponent
+import com.makeevrserg.empireprojekt.mobile.wear.features.status.ui.StatusesScreen
 import com.makeevrserg.empireprojekt.mobile.wear.ui.SwipeToDismissBox
 
 @Composable
 fun RootScreen(
     rootComponent: RootComponent,
 ) {
-//    val child by rootComponent.childStack.subscribeAsState()
-//    val instance = child.active.instance
     SwipeToDismissBox(stack = rootComponent.childStack, onDismissed = rootComponent::pop) { child ->
         when (val instance = child.instance) {
             is RootComponent.Child.Main -> MainScreen(
