+++ instances/wearApp/src/main/java/com/makeevrserg/empireprojekt/mobile/wear/features/status/ui/preview/StatusesScreenPreview.kt
@@ -1,10 +1,10 @@
-package com.makeevrserg.empireprojekt.mobile.wear.features.status.preview
+package com.makeevrserg.empireprojekt.mobile.wear.features.status.ui.preview
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import com.makeevrserg.empireprojekt.mobile.features.theme.ApplicationTheme
-import com.makeevrserg.empireprojekt.mobile.wear.features.status.StatusesScreen
-import com.makeevrserg.empireprojekt.mobile.wear.features.status.WearStatusComponent
+import com.makeevrserg.empireprojekt.mobile.wear.features.status.presentation.WearStatusComponent
+import com.makeevrserg.empireprojekt.mobile.wear.features.status.ui.StatusesScreen
 
 @Preview
 @Composable
